requirements.txt
annotated-types==0.7.0
bcrypt==4.2.0
blinker==1.8.2
bson==0.5.10
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
Werkzeug==3.0.4
